Split out getTime into a separate file
